+++ api/pkg/handler/test/fake_auth.go
@@ -1,4 +1,4 @@
-package fake
+package test
 
 import (
 	"context"
